Move 8 mock-based test files from tests/ to inline #[cfg(test)] modules in src/
Mock-based component tests for serial managers, device traits, client
mocks, connection mocks, process mocks, and FITS utilities were sitting
in tests/ as separate binaries. Each independently compiled and linked
against the full dependency tree.

Moving them into #[cfg(test)] modules in their corresponding src/ files
eliminates 8 more test binaries (from 42 down to 34). The mock structs
are kept per-module to maintain self-containment.

Services affected: ppba-driver (3 files), qhy-focuser (1 file),
phd2-guider (4 files).
