+++ docs/services/ppba-driver.md
@@ -203,11 +203,8 @@ ppba-driver/
 │   │   ├── sensor_readings.feature
 │   │   ├── observing_conditions.feature
 │   │   └── server_registration.feature
-│   ├── test_serial_manager.rs        # SerialManager internal API unit tests
-│   ├── test_switch_device.rs         # Switch device error mapping unit tests
-│   ├── test_oc_device.rs             # ObservingConditions device unit tests
 │   └── conformu_integration.rs       # ASCOM ConformU compliance tests
-│   # Unit tests for protocol, config, error, switches are in src/ as #[cfg(test)] modules
+│   # Unit and mock-based tests are in src/ as #[cfg(test)] modules
 └── examples/
     ├── config-linux.json
     ├── config-macos.json
